Get rid of JavaScript banzai on settings page
Just load regular HTML pages; it grew like this with just one page, and then I
was like "let's add a tab", and then two, and then three, and then it got out of
control.

It never worked well: it had "flashes" of the full page before the JS could
render, linking directly to a setting element was always hard, and the JS code
was far too complex.

Just go to a new page for every tab. It worked great in 1997, and it still works
great in 2020.
