+++ handlers/backend.go
@@ -5,15 +5,10 @@
 package handlers
 
 import (
-	"compress/gzip"
 	"context"
 	"fmt"
-	"io"
 	"net"
 	"net/http"
-	"net/url"
-	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"sync"
@@ -23,24 +18,17 @@ import (
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
 	"github.com/monoculum/formam"
-	"zgo.at/blackmail"
-	"zgo.at/errors"
 	"zgo.at/goatcounter"
-	"zgo.at/goatcounter/acme"
-	"zgo.at/goatcounter/bgrun"
 	"zgo.at/goatcounter/cfg"
 	"zgo.at/goatcounter/pack"
-	"zgo.at/goatcounter/widgets"
 	"zgo.at/guru"
 	"zgo.at/isbot"
-	"zgo.at/tz"
 	"zgo.at/zdb"
 	"zgo.at/zhttp"
 	"zgo.at/zhttp/header"
 	"zgo.at/zhttp/mware"
 	"zgo.at/zhttp/ztpl"
 	"zgo.at/zlog"
-	"zgo.at/zstd/zcrypto"
 	"zgo.at/zstd/zint"
 	"zgo.at/zstd/zjson"
 	"zgo.at/zstd/zstring"
@@ -151,28 +139,9 @@ func (h backend) Mount(r chi.Router, db zdb.DB) {
 				billing{}.mount(a, af)
 			}
 			af.Get("/updates", zhttp.Wrap(h.updates))
-			af.Get("/settings", zhttp.Wrap(h.settings))
-			af.Get("/code", zhttp.Wrap(h.code))
-			af.Get("/ip", zhttp.Wrap(h.ip))
-			af.Post("/save-settings", zhttp.Wrap(h.saveSettings))
-			af.Post("/save-widgets", zhttp.Wrap(h.saveWidgets))
-			af.Post("/save-view", zhttp.Wrap(h.saveView))
-			af.Get("/settings/change-code", zhttp.Wrap(h.changeCode))
-			af.Post("/settings/change-code", zhttp.Wrap(h.changeCode))
-			af.With(mware.Ratelimit(mware.RatelimitOptions{
-				Client:  mware.RatelimitIP,
-				Store:   mware.NewRatelimitMemory(),
-				Limit:   mware.RatelimitLimit(1, 3600),
-				Message: "you can request only one export per hour",
-			})).Post("/export", zhttp.Wrap(h.startExport))
-			af.Get("/export/{id}", zhttp.Wrap(h.downloadExport))
-			af.Post("/import", zhttp.Wrap(h.importFile))
-			af.Post("/add", zhttp.Wrap(h.addSubsite))
-			af.Get("/remove/{id}", zhttp.Wrap(h.removeSubsiteConfirm))
-			af.Post("/remove/{id}", zhttp.Wrap(h.removeSubsite))
-			af.Get("/purge", zhttp.Wrap(h.purgeConfirm))
-			af.Post("/purge", zhttp.Wrap(h.purge))
-			af.Post("/delete", zhttp.Wrap(h.delete))
+			r.Get("/code", zhttp.Wrap(h.code))
+
+			settings{}.mount(af)
 			admin{}.mount(af, db)
 		}
 	}
@@ -554,48 +523,6 @@ func (h backend) updates(w http.ResponseWriter, r *http.Request) error {
 	}{newGlobals(w, r), up, seenat})
 }
 
-func (h backend) settings(w http.ResponseWriter, r *http.Request) error {
-	return h.settingsTpl(w, r, nil, "")
-}
-
-func (h backend) settingsTpl(w http.ResponseWriter, r *http.Request, verr *zvalidate.Validator, loadTab string) error {
-	var sites goatcounter.Sites
-	err := sites.ListSubs(r.Context())
-	if err != nil {
-		return err
-	}
-
-	var exports goatcounter.Exports
-	err = exports.List(r.Context())
-	if err != nil {
-		return err
-	}
-
-	var tokens goatcounter.APITokens
-	err = tokens.List(r.Context())
-	if err != nil {
-		return err
-	}
-
-	wid := widgets.FromSiteWidgets(Site(r.Context()).Settings.Widgets, widgets.FilterInternal)
-	del := map[string]interface{}{
-		"ContactMe": r.URL.Query().Get("contact_me") == "true",
-		"Reason":    r.URL.Query().Get("reason"),
-	}
-
-	return zhttp.Template(w, "backend_settings.gohtml", struct {
-		Globals
-		SubSites  goatcounter.Sites
-		Validate  *zvalidate.Validator
-		Timezones []*tz.Zone
-		Delete    map[string]interface{}
-		Exports   goatcounter.Exports
-		APITokens goatcounter.APITokens
-		Widgets   widgets.List
-		LoadTab   string
-	}{newGlobals(w, r), sites, verr, tz.Zones, del, exports, tokens, wid, loadTab})
-}
-
 func (h backend) code(w http.ResponseWriter, r *http.Request) error {
 	var sites goatcounter.Sites
 	err := sites.ListSubs(r.Context())
@@ -616,471 +543,6 @@ func (h backend) code(w http.ResponseWriter, r *http.Request) error {
 		SubSites    goatcounter.Sites
 		CountDomain string
 	}{newGlobals(w, r), sites, cd})
-}
-
-func (h backend) ip(w http.ResponseWriter, r *http.Request) error {
-	return zhttp.String(w, r.RemoteAddr)
-}
-
-func (h backend) saveSettings(w http.ResponseWriter, r *http.Request) error {
-	v := zvalidate.New()
-
-	args := struct {
-		Cname      string                   `json:"cname"`
-		LinkDomain string                   `json:"link_domain"`
-		Settings   goatcounter.SiteSettings `json:"settings"`
-		User       goatcounter.User         `json:"user"`
-	}{}
-	_, err := zhttp.Decode(r, &args)
-	if err != nil {
-		ferr, ok := err.(*formam.Error)
-		if !ok || ferr.Code() != formam.ErrCodeConversion {
-			return err
-		}
-		v.Append(ferr.Path(), "must be a number")
-
-		// TODO: we return here because formam stops decoding on the first
-		// error. We should really fix this in formam, but it's an incompatible
-		// change.
-		return h.settingsTpl(w, r, &v, "")
-	}
-
-	txctx, tx, err := zdb.Begin(r.Context())
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback()
-
-	user := goatcounter.GetUser(txctx)
-
-	emailChanged := false
-	if cfg.GoatcounterCom && args.User.Email != user.Email {
-		emailChanged = true
-	}
-
-	user.Email = args.User.Email
-	err = user.Update(txctx, emailChanged)
-	if err != nil {
-		var vErr *zvalidate.Validator
-		if !errors.As(err, &vErr) {
-			return err
-		}
-		v.Sub("user", "", err)
-	}
-
-	site := Site(txctx)
-	site.Settings = args.Settings
-	site.LinkDomain = args.LinkDomain
-	if args.Cname != "" && !site.PlanCustomDomain(txctx) {
-		return guru.New(http.StatusForbidden, "need a business plan to set custom domain")
-	}
-
-	makecert := false
-	if args.Cname == "" {
-		site.Cname = nil
-	} else {
-		if site.Cname == nil || *site.Cname != args.Cname {
-			makecert = true // Make after we persisted to DB.
-		}
-		site.Cname = &args.Cname
-	}
-
-	err = site.Update(txctx)
-	if err != nil {
-		var vErr *zvalidate.Validator
-		if !errors.As(err, &vErr) {
-			return err
-		}
-		v.Sub("site", "", err)
-	}
-
-	if v.HasErrors() {
-		return h.settingsTpl(w, r, &v, "")
-	}
-
-	err = tx.Commit()
-	if err != nil {
-		return err
-	}
-
-	if emailChanged {
-		sendEmailVerify(site, user)
-	}
-
-	if makecert {
-		ctx := goatcounter.NewContext(r.Context())
-		bgrun.Run(fmt.Sprintf("acme.Make:%s", args.Cname), func() {
-			err := acme.Make(args.Cname)
-			if err != nil {
-				zlog.Field("domain", args.Cname).Error(err)
-				return
-			}
-
-			err = site.UpdateCnameSetupAt(ctx)
-			if err != nil {
-				zlog.Field("domain", args.Cname).Error(err)
-			}
-		})
-	}
-
-	zhttp.Flash(w, "Saved!")
-	return zhttp.SeeOther(w, "/settings")
-}
-
-func (h backend) saveWidgets(w http.ResponseWriter, r *http.Request) error {
-	err := r.ParseForm()
-	if err != nil {
-		return err
-	}
-
-	site := Site(r.Context())
-
-	if r.Form.Get("reset") != "" {
-		site.Settings.Widgets = nil
-		site.Defaults(r.Context())
-		err = site.Update(r.Context())
-		if err != nil {
-			return err
-		}
-
-		zhttp.Flash(w, "Reset to defaults!")
-		return zhttp.SeeOther(w, "/settings#tab-dashboard")
-	}
-
-	parse := make(map[string]map[string]string)
-	for k, v := range r.Form {
-		if !strings.HasPrefix(k, "widgets.") {
-			continue
-		}
-		k = k[8:]
-
-		name, key := zstring.Split2(k, ".")
-		if parse[name] == nil {
-			parse[name] = make(map[string]string)
-		}
-		parse[name][key] = v[0]
-	}
-
-	site.Settings.Widgets = make(goatcounter.Widgets, len(parse))
-	for k, v := range parse {
-		pos, err := strconv.Atoi(v["index"])
-		if err != nil {
-			return err
-		}
-
-		w := goatcounter.Widget{"name": k, "on": v["on"] == "on"}
-		for sName, sVal := range v {
-			if strings.HasPrefix(sName, "s.") {
-				err := w.SetSetting(k, sName[2:], sVal)
-				if err != nil {
-					return err
-				}
-			}
-		}
-		site.Settings.Widgets[pos] = w
-	}
-
-	err = site.Update(r.Context())
-	if err != nil {
-		var v *zvalidate.Validator
-		if errors.As(err, &v) {
-			return h.settingsTpl(w, r, v, "dashboard")
-		}
-		return err
-	}
-
-	zhttp.Flash(w, "Saved!")
-	return zhttp.SeeOther(w, "/settings#tab-dashboard")
-}
-
-func (h backend) saveView(w http.ResponseWriter, r *http.Request) error {
-	site := Site(r.Context())
-	v, i := site.Settings.Views.Get("default") // TODO: only default view for now.
-	_, err := zhttp.Decode(r, &v)
-	if err != nil {
-		return err
-	}
-
-	site.Settings.Views[i] = v
-	err = site.Update(r.Context())
-	if err != nil {
-		return err
-	}
-
-	return zhttp.JSON(w, map[string]string{})
-}
-
-func (h backend) changeCode(w http.ResponseWriter, r *http.Request) error {
-	if r.Method == "GET" {
-		return zhttp.Template(w, "backend_settings_code.gohtml", struct {
-			Globals
-		}{newGlobals(w, r)})
-	}
-
-	var args struct {
-		Code string `json:"code"`
-	}
-	_, err := zhttp.Decode(r, &args)
-	if err != nil {
-		return err
-	}
-
-	site := Site(r.Context())
-	err = site.UpdateCode(r.Context(), args.Code)
-	if err != nil {
-		return err
-	}
-
-	zhttp.Flash(w, "Saved!")
-	return zhttp.SeeOther(w, site.URL()+"/settings")
-}
-
-func (h backend) importFile(w http.ResponseWriter, r *http.Request) error {
-	v := zvalidate.New()
-	replace := v.Boolean("replace", r.Form.Get("replace"))
-	if v.HasErrors() {
-		return v
-	}
-
-	file, head, err := r.FormFile("csv")
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	var fp io.ReadCloser = file
-	if strings.HasSuffix(head.Filename, ".gz") {
-		fp, err = gzip.NewReader(file)
-		if err != nil {
-			return guru.Errorf(400, "could not read as gzip: %w", err)
-		}
-	}
-	defer fp.Close()
-
-	user := goatcounter.GetUser(r.Context())
-	ctx := goatcounter.NewContext(r.Context())
-	n := 0
-	bgrun.Run(fmt.Sprintf("import:%d", Site(ctx).ID), func() {
-		firstHitAt, err := goatcounter.Import(ctx, fp, replace, true, func(hit goatcounter.Hit, final bool) {
-			if final {
-				return
-			}
-
-			goatcounter.Memstore.Append(hit)
-			n++
-
-			// Spread out the load a bit.
-			if n%5000 == 0 {
-				goatcounter.PersistRunner.Run <- struct{}{}
-				for bgrun.Running("cron:PersistAndStat") {
-					time.Sleep(250 * time.Millisecond)
-				}
-			}
-		})
-		if err != nil {
-			if e, ok := err.(*errors.StackErr); ok {
-				err = e.Unwrap()
-			}
-
-			sendErr := blackmail.Send("GoatCounter import error",
-				blackmail.From("GoatCounter import", cfg.EmailFrom),
-				blackmail.To(user.Email),
-				blackmail.BodyMustText(goatcounter.EmailTemplate("email_import_error.gotxt", struct {
-					Error error
-				}{err})))
-			if sendErr != nil {
-				zlog.Error(sendErr)
-			}
-		}
-
-		if firstHitAt != nil {
-			err := Site(ctx).UpdateFirstHitAt(ctx, *firstHitAt)
-			if err != nil {
-				zlog.Error(err)
-			}
-		}
-	})
-
-	zhttp.Flash(w, "Import started in the background; you’ll get an email when it’s done.")
-	return zhttp.SeeOther(w, "/settings#tab-export")
-}
-
-func (h backend) startExport(w http.ResponseWriter, r *http.Request) error {
-	r.ParseForm()
-
-	v := zvalidate.New()
-	startFrom := v.Integer("startFrom", r.Form.Get("startFrom"))
-	if v.HasErrors() {
-		return v
-	}
-
-	var export goatcounter.Export
-	fp, err := export.Create(r.Context(), startFrom)
-	if err != nil {
-		return err
-	}
-
-	ctx := goatcounter.NewContext(r.Context())
-	bgrun.Run(fmt.Sprintf("export web:%d", Site(ctx).ID),
-		func() { export.Run(ctx, fp, true) })
-
-	zhttp.Flash(w, "Export started in the background; you’ll get an email with a download link when it’s done.")
-	return zhttp.SeeOther(w, "/settings#tab-export")
-}
-
-func (h backend) downloadExport(w http.ResponseWriter, r *http.Request) error {
-	v := zvalidate.New()
-	id := v.Integer("id", chi.URLParam(r, "id"))
-	if v.HasErrors() {
-		return v
-	}
-
-	var export goatcounter.Export
-	err := export.ByID(r.Context(), id)
-	if err != nil {
-		return err
-	}
-
-	fp, err := os.Open(export.Path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			zhttp.FlashError(w, "It looks like there is no export yet.")
-			return zhttp.SeeOther(w, "/settings#tab-export")
-		}
-
-		return err
-	}
-	defer fp.Close()
-
-	err = header.SetContentDisposition(w.Header(), header.DispositionArgs{
-		Type:     header.TypeAttachment,
-		Filename: filepath.Base(export.Path),
-	})
-	if err != nil {
-		return err
-	}
-
-	w.Header().Set("Content-Type", "application/gzip")
-	return zhttp.Stream(w, fp)
-}
-
-func (h backend) removeSubsiteConfirm(w http.ResponseWriter, r *http.Request) error {
-	v := zvalidate.New()
-	id := v.Integer("id", chi.URLParam(r, "id"))
-	if v.HasErrors() {
-		return v
-	}
-
-	var s goatcounter.Site
-	err := s.ByID(r.Context(), id)
-	if err != nil {
-		return err
-	}
-
-	return zhttp.Template(w, "backend_remove.gohtml", struct {
-		Globals
-		Site goatcounter.Site
-	}{newGlobals(w, r), s})
-}
-
-func (h backend) removeSubsite(w http.ResponseWriter, r *http.Request) error {
-	v := zvalidate.New()
-	id := v.Integer("id", chi.URLParam(r, "id"))
-	if v.HasErrors() {
-		return v
-	}
-
-	var s goatcounter.Site
-	err := s.ByID(r.Context(), id)
-	if err != nil {
-		return err
-	}
-
-	err = s.Delete(r.Context())
-	if err != nil {
-		return err
-	}
-
-	zhttp.Flash(w, "Site ‘%s ’removed.", s.URL())
-	return zhttp.SeeOther(w, "/settings#tab-sites")
-}
-
-func (h backend) addSubsite(w http.ResponseWriter, r *http.Request) error {
-	var args struct {
-		Code  string `json:"code"`
-		Cname string `json:"cname"`
-	}
-	_, err := zhttp.Decode(r, &args)
-	if err != nil {
-		return err
-	}
-
-	parent := Site(r.Context())
-	site := goatcounter.Site{
-		Parent:   &parent.ID,
-		Plan:     goatcounter.PlanChild,
-		Settings: parent.Settings,
-	}
-	if cfg.GoatcounterCom {
-		site.Code = args.Code
-	} else {
-		site.Code = "serve-" + zcrypto.Secret64()
-		site.Cname = &args.Cname
-	}
-
-	err = zdb.TX(r.Context(), func(ctx context.Context, tx zdb.DB) error {
-		err := site.Insert(ctx)
-		if err != nil {
-			return err
-		}
-		if !cfg.GoatcounterCom {
-			return site.UpdateCnameSetupAt(ctx)
-		}
-		return nil
-	})
-	if err != nil {
-		zhttp.FlashError(w, err.Error())
-		return zhttp.SeeOther(w, "/settings#tab-sites")
-	}
-
-	zhttp.Flash(w, "Site ‘%s’ added.", site.URL())
-	return zhttp.SeeOther(w, "/settings#tab-sites")
-}
-
-func (h backend) purgeConfirm(w http.ResponseWriter, r *http.Request) error {
-	path := strings.TrimSpace(r.URL.Query().Get("path"))
-	title := r.URL.Query().Get("match-title") == "on"
-
-	var list goatcounter.HitStats
-	err := list.ListPathsLike(r.Context(), path, title)
-	if err != nil {
-		return err
-	}
-
-	return zhttp.Template(w, "backend_purge.gohtml", struct {
-		Globals
-		PurgePath string
-		List      goatcounter.HitStats
-	}{newGlobals(w, r), path, list})
-}
-
-func (h backend) purge(w http.ResponseWriter, r *http.Request) error {
-	paths, err := zint.Split(r.Form.Get("paths"), ",")
-	if err != nil {
-		return err
-	}
-
-	ctx := goatcounter.NewContext(r.Context())
-	bgrun.Run(fmt.Sprintf("purge:%d", Site(ctx).ID), func() {
-		var list goatcounter.Hits
-		err := list.Purge(ctx, paths)
-		if err != nil {
-			zlog.Error(err)
-		}
-	})
-
-	zhttp.Flash(w, "Started in the background; may take about 10-20 seconds to fully process.")
-	return zhttp.SeeOther(w, "/settings#tab-purge")
 }
 
 func hasPlan(site *goatcounter.Site) (bool, error) {
@@ -1115,73 +577,6 @@ func hasPlan(site *goatcounter.Site) (bool, error) {
 	}
 
 	return true, nil
-}
-
-func (h backend) delete(w http.ResponseWriter, r *http.Request) error {
-	site := Site(r.Context())
-
-	if cfg.GoatcounterCom {
-		var args struct {
-			Reason    string `json:"reason"`
-			ContactMe bool   `json:"contact_me"`
-		}
-		_, err := zhttp.Decode(r, &args)
-		if err != nil {
-			zlog.Error(err)
-		}
-
-		has, err := hasPlan(site)
-		if err != nil {
-			return err
-		}
-		if has {
-			zhttp.FlashError(w, "This site still has a Stripe subscription; cancel that first on the billing page.")
-			q := url.Values{}
-			q.Set("reason", args.Reason)
-			q.Set("contact_me", fmt.Sprintf("%t", args.ContactMe))
-			return zhttp.SeeOther(w, "/settings?"+q.Encode()+"#tab-delete")
-		}
-
-		if args.Reason != "" {
-			bgrun.Run("email:deletion", func() {
-				contact := "false"
-				if args.ContactMe {
-					var u goatcounter.User
-					err := u.BySite(r.Context(), site.ID)
-					if err != nil {
-						zlog.Error(err)
-					} else {
-						contact = u.Email
-					}
-				}
-
-				blackmail.Send("GoatCounter deletion",
-					blackmail.From("GoatCounter deletion", cfg.EmailFrom),
-					blackmail.To(cfg.EmailFrom),
-					blackmail.Bodyf(`Deleted: %s (%d): contact_me: %s; reason: %s`,
-						site.Code, site.ID, contact, args.Reason))
-			})
-		}
-	}
-
-	err := site.Delete(r.Context())
-	if err != nil {
-		return err
-	}
-
-	if site.Parent != nil {
-		var p goatcounter.Site
-		err := p.ByID(r.Context(), *site.Parent)
-		if err != nil {
-			return err
-		}
-		return zhttp.SeeOther(w, p.URL())
-	}
-
-	if cfg.GoatcounterCom {
-		return zhttp.SeeOther(w, "https://"+cfg.Domain)
-	}
-	return zhttp.SeeOther(w, "/")
 }
 
 func getPeriod(w http.ResponseWriter, r *http.Request, site *goatcounter.Site) (time.Time, time.Time, error) {
