+++ handlers/backend_test.go
@@ -7,7 +7,6 @@ package handlers
 import (
 	"context"
 	"net/http"
-	"net/http/httptest"
 	"net/url"
 	"sort"
 	"testing"
@@ -15,7 +14,6 @@ import (
 
 	"github.com/go-chi/chi"
 	"zgo.at/goatcounter"
-	"zgo.at/goatcounter/bgrun"
 	"zgo.at/goatcounter/cfg"
 	"zgo.at/goatcounter/gctest"
 	"zgo.at/isbot"
@@ -322,86 +320,6 @@ func TestBackendCountSessions(t *testing.T) {
 	hits2 = checkHits(ctx2, 4)
 	want = []int{1, 1, 2, 3, 3, 1, 2, 1, 3, 4, 5}
 	checkSess(append(hits1, hits2...), want)
-}
-
-func TestBackendTpl(t *testing.T) {
-	tests := []handlerTest{
-		{
-			setup: func(ctx context.Context, t *testing.T) {
-				now := time.Date(2019, 8, 31, 14, 42, 0, 0, time.UTC)
-				gctest.StoreHits(ctx, t, false, []goatcounter.Hit{
-					{Site: 1, Path: "/asd", Title: "AAA", CreatedAt: now},
-					{Site: 1, Path: "/asd", Title: "AAA", CreatedAt: now},
-					{Site: 1, Path: "/zxc", Title: "BBB", CreatedAt: now},
-				}...)
-			},
-			router:   newBackend,
-			path:     "/purge?path=/asd",
-			auth:     true,
-			wantCode: 200,
-			wantBody: "<tr><td>2</td><td>/asd</td><td>AAA</td></tr>",
-		},
-
-		{
-			setup: func(ctx context.Context, t *testing.T) {
-				one := int64(1)
-				ss := goatcounter.Site{
-					Code:   "subsite",
-					Parent: &one,
-					Plan:   goatcounter.PlanChild,
-				}
-				err := ss.Insert(ctx)
-				if err != nil {
-					panic(err)
-				}
-			},
-			router:   newBackend,
-			path:     "/remove/2",
-			auth:     true,
-			wantCode: 200,
-			wantBody: "Are you sure you want to remove the site",
-		},
-	}
-
-	for _, tt := range tests {
-		runTest(t, tt, nil)
-	}
-}
-
-func TestBackendPurge(t *testing.T) {
-	tests := []handlerTest{
-		{
-			setup: func(ctx context.Context, t *testing.T) {
-				now := time.Date(2019, 8, 31, 14, 42, 0, 0, time.UTC)
-				gctest.StoreHits(ctx, t, false, []goatcounter.Hit{
-					{Site: 1, Path: "/asd", CreatedAt: now},
-					{Site: 1, Path: "/asd", CreatedAt: now},
-					{Site: 1, Path: "/zxc", CreatedAt: now},
-				}...)
-			},
-			router:       newBackend,
-			path:         "/purge",
-			body:         map[string]string{"path": "/asd", "paths": "1,"},
-			method:       "POST",
-			auth:         true,
-			wantFormCode: 303,
-		},
-	}
-
-	for _, tt := range tests {
-		runTest(t, tt, func(t *testing.T, rr *httptest.ResponseRecorder, r *http.Request) {
-			bgrun.Wait()
-
-			var hits goatcounter.Hits
-			err := hits.TestList(r.Context(), false)
-			if err != nil {
-				t.Fatal(err)
-			}
-			if len(hits) != 1 {
-				t.Errorf("%d hits in DB; expected 1:\n%v", len(hits), hits)
-			}
-		})
-	}
 }
 
 func BenchmarkCount(b *testing.B) {
