+++ tpl/_backend_sitecode.gohtml
@@ -47,7 +47,7 @@ image. The easiest way to do this is from the JavaScript integration:</p>
 
 {{if .Site.ID}}
 <strong>Note</strong>: you will need to enable “Allow adding visitor counts on your website”
-in your <a href="/settings#tab-setting">site settings</a>; this defaults to
+in your <a href="/settings/main">site settings</a>; this defaults to
 <em>off</em> to prevent accidental unintentional leaking of data.
 {{else}}
 <strong>Note</strong>: you will need to enable “Allow adding visitor counts on your website”
