+++ tpl/dashboard.gohtml
@@ -46,7 +46,7 @@
 				{{/* TODO: it might be better to load the settings page "inline"
 				here, instead of a settings tab; would also declutter that a bit
 				since we can remove it there. */}}
-				<a href="/settings#tab-dashboard">Configure dashboard layout</a><br>
+				<a href="/settings/dashboard">Configure dashboard layout</a><br>
 				<small>Change what to display on the dashboard and in what order.</small>
 			</div>
 		</div>
