+++ tpl/settings_dashboard.gohtml
@@ -1,0 +1,45 @@
+{{template "_backend_top.gohtml" .}}
+
+{{template "_settings_nav.gohtml" .}}
+
+<h2 id="dashboard">Dashboard</h2>
+
+<script crossorigin="anonymous" src="{{.Static}}/dragula.js?v={{.Version}}"></script>
+<form method="post" action="/settings/widgets" id="widget-settings">
+	<input type="hidden" name="csrf" value="{{.User.CSRFToken}}">
+	<input type="hidden" name="reset" value="">
+
+	{{range $i, $w := .Widgets}}
+		<div class="widget" id="widget-{{$w.Name}}">
+			<span class="drag-handle" title="Drag to reorder"></span>
+			<input type="hidden" name="widgets.{{$w.Name}}.index" value="{{$i}}" class="index">
+			<label class="main">
+				<input type="checkbox" name="widgets.{{$w.Name}}.on" {{if $.Site.Settings.Widgets.On $w.Name}}checked{{end}}>
+				<strong>{{$w.Label}}</strong>
+			</label>
+			<input type="hidden" name="widgets.{{$w.Name}}.on" value="off">
+
+			<div class="widget-settings">
+				{{range $k, $v := $.Site.Settings.Widgets.GetSettings $w.Name}}
+					{{ $id := (print "widgets_" $w.Name "_s_" $k) }}
+					{{ $n  := (print "widgets." $w.Name ".s." $k) }}
+					{{if eq $v.Type "checkbox"}}
+						<label>{{checkbox $v.Value $n}} {{$v.Label}}</label>
+					{{else}}
+						<label for="{{$id}}">{{$v.Label}}</label>
+						<input type="{{$v.Type}}" name="{{$n}}" id="{{$id}}" value="{{$v.Value}}">
+					{{end}}
+					<small class="help help-{{$v.Type}}">{{$v.Help}}</small>
+					{{validate (print "widgets." $w.Name ".s." $k) $.Validate}}
+					<br>
+				{{end}}
+			</div>
+		</div>
+	{{end}}
+	<div class="widget-save">
+		<button type="save">Save</button>
+		<button class="widgets-reset">Reset to defaults</button>
+	</div>
+</form>
+
+{{template "_backend_bottom.gohtml" .}}
