+++ tpl/settings_main.gohtml
@@ -1,0 +1,165 @@
+{{template "_backend_top.gohtml" .}}
+
+{{template "_settings_nav.gohtml" .}}
+
+<h2 id="setting">Settings</h2>
+
+<div class="form-wrap">
+	<form method="post" action="/settings/main" class="vertical">
+		<input type="hidden" name="csrf" value="{{.User.CSRFToken}}">
+
+		<fieldset id="section-site">
+			<legend>Site settings</legend>
+			<label for="link_domain">Your site</label>
+			<input type="text" name="link_domain" id="link_domain" value="{{.Site.LinkDomain}}">
+			{{validate "site.link_domain" .Validate}}
+			<span>Your site’s domain, e.g. <em>“www.example.com”</em>, used for linking to the page in the overview.</span>
+
+			<label for="user.email">Your email</label>
+			<input type="text" name="user.email" id="user.email" value="{{.User.Email}}">
+			{{validate "user.email" .Validate}}
+			<span>You will need to re-verify the new address if you change it.</span>
+
+			{{if .GoatcounterCom}}
+				<label>{{checkbox .Site.Settings.AllowAdmin "settings.allow_admin"}}
+					Allow admin access</label>
+				<span>Normally administrators can’t “log in” to your site, but
+					if this is enabled they can. You may be asked to enable this
+					for support requests.</span>
+			{{end}}
+
+			<label>{{checkbox .Site.Settings.Public "settings.public"}}
+				Make statistics publicly viewable</label>
+			<span>Anyone can view the statistics without logging in.</span>
+
+			<label>{{checkbox .Site.Settings.AllowCounter "settings.allow_counter"}}
+				Allow adding visitor counts on your website</label>
+			<span>See <a href="/code#visitor-counter">the documentation</a> for details on how to use.</span>
+		</fieldset>
+
+		<fieldset id="section-domain">
+			<legend>Domain settings</legend>
+
+			{{if .GoatcounterCom}}
+				<label for="code">Code</label>
+				<input type="text" disabled id="code" class="inline" value="{{.Site.Code}}">
+				<span class="help">You will access your account at https://<em>[my-code]</em>.{{.Domain}}.
+				<a href="/settings/change-code">Change</a>.</span>
+			{{end}}
+
+			{{if .GoatcounterCom}}
+				<label for="cname">Custom domain</label>
+				<input type="text" name="cname" id="cname" value="{{if .Site.Cname}}{{.Site.Cname}}{{end}}"
+					{{if not (.Site.PlanCustomDomain .Context)}}disabled{{end}}>
+				<span>Custom domain, e.g. <em>“stats.example.com”</em>.
+					{{if not (.Site.PlanCustomDomain .Context)}}
+						Requires Personal Plus or Business plan (you’re
+						on the {{.Site.Plan}} plan; see
+						<a href="/billing">billing</a>).
+					{{else}}
+						{{if .Site.CnameSetupAt}}
+							<br>Domain verified and set up (note: it may take up to an hour for the certificate to work).
+						{{else if .Site.Cname}}
+							<br><span style="color: red;">Not yet verified</span>; set a CNAME record to
+							<code>{{.Site.Code}}.{{.Domain}}</code>.
+							<a href="https://www.goatcounter.com/help#custom-domain" target="_blank">Detailed instructions</a>.
+							Verification runs every 2 hours.
+						{{else}}
+							Set a CNAME record to
+							<code>{{.Site.Code}}.{{.Domain}}</code>.
+							<a href="https://www.goatcounter.com/help#custom-domain" target="_blank">Detailed instructions</a>.
+						{{end}}
+					{{end}}</span>
+			{{else}}
+				<label for="cname">GoatCounter domain</label>
+				<input type="text" name="cname" id="cname" value="{{if .Site.Cname}}{{.Site.Cname}}{{end}}">
+				<span>You GoatCounter installation’s domain, e.g. <em>“stats.example.com”</em>.</span>
+			{{end}}
+		</fieldset>
+
+		{{/*
+		<fieldset id="section-user-pref">
+			<legend>User info and preferences</legend>
+		</fieldset>
+		*/}}
+
+		<fieldset id="section-l10n">
+			<legend>Localisation preferences</legend>
+
+			<label for="date_format">Date format</label>
+			<select name="settings.date_format" id="date_format">
+				<option {{option_value .Site.Settings.DateFormat "2006-01-02"}}>year-month-day (2006-01-02)</option>
+				<option {{option_value .Site.Settings.DateFormat "02-01-2006"}}>day-month-year (02-01-2006)</option>
+				<option {{option_value .Site.Settings.DateFormat "01/02/06"}}>month/day/year (01/02/06)</option>
+				<option {{option_value .Site.Settings.DateFormat "2 Jan ’06"}}>Short text (2 Jan '06)</option>
+				<option {{option_value .Site.Settings.DateFormat "Mon Jan 2 2006"}}>Long text (Mon Jan 2 2006)</option>
+			</select>
+
+			<label>{{checkbox .Site.Settings.TwentyFourHours "settings.twenty_four_hours"}}
+				24-hour clock (13:00)</label>
+
+			<label>{{checkbox .Site.Settings.SundayStartsWeek "settings.sunday_starts_week"}}
+				Week starts on Sunday</label>
+
+			<label for="number_format">Thousands separator</label>
+			<select name="settings.number_format" id="number_format">
+				<option {{option_value (string .Site.Settings.NumberFormat) "8239"}}>Thin space (42 123)</option>
+				<option {{option_value (string .Site.Settings.NumberFormat) "160"}}>Space (42 123)</option>
+				<option {{option_value (string .Site.Settings.NumberFormat) "44"}}>Comma (42,123)</option>
+				<option {{option_value (string .Site.Settings.NumberFormat) "46"}}>Dot (42.123)</option>
+				<option {{option_value (string .Site.Settings.NumberFormat) "39"}}>Apostrophe (42'123)</option>
+				<option {{option_value (string .Site.Settings.NumberFormat) "1"}}>None (42123)</option>
+			</select>
+			{{validate "site.settings.number_format" .Validate}}
+
+			<label for="timezone">Timezone</label>
+			<select name="settings.timezone" id="timezone">
+				<option {{option_value $.Site.Settings.Timezone.String ".UTC"}}>UTC</option>
+				{{range $tz := .Timezones}}<option {{option_value $.Site.Settings.Timezone.String $tz.String}}>{{$tz.Display}}</option>
+				{{end}}
+			</select>
+			{{validate "site.settings.timezone" .Validate}}
+			<span><a href="#_" id="set-local-tz">Set from browser</a></span>
+		</fieldset>
+
+		<fieldset id="section-tracking">
+			<legend>Tracking</legend>
+
+			<label for="data_retention">Data retention in days</label>
+			<input type="number" name="settings.data_retention" id="limits_page" value="{{.Site.Settings.DataRetention}}">
+			{{validate "site.settings.data_retention" .Validate}}
+			<span class="help">Pageviews and all associated data will be permanently removed after this many days. Set to <code>0</code> to never delete.</span>
+
+			<label>Ignore IPs</label>
+			<input type="text" name="settings.ignore_ips" value="{{.Site.Settings.IgnoreIPs}}">
+			{{validate "site.settings.ignore_ips" .Validate}}
+			<span>Never count requests coming from these IP addresses.
+				Comma-separated. Only supports exact matches.
+				<a href="#_" id="add-ip">Add your current IP</a>.
+				{{if .Site.LinkDomain}}<br>
+				Alternatively, <a href="http://{{.Site.LinkDomain}}#toggle-goatcounter">disable for this browser</a> (click again to enable).{{end}}
+			</span>
+
+			<label>Campaign parameters</label>
+			<input type="text" name="settings.campaigns" value="{{.Site.Settings.Campaigns}}">
+			{{validate "site.settings.campaigns" .Validate}}
+			<span>
+				List of parameters to count as ‘campaigns’; if set then the
+				value will be set as the referrer, overriding any Referer
+				header.{{/* <a href="/code#campaigns">Details</a>.
+				Comma-separated; first match takes precedence.*/}}
+			</span>
+		</fieldset>
+
+		<div class="flex-break"></div>
+		<button type="submit">Save</button>
+	</form>
+
+	{{if has_errors .Validate}}
+		<div class="flash flash-e"
+			style="position: fixed; bottom: 0; right: 0; min-width: 20em; z-index: 5; text-align: left;">
+		Additional errors:{{.Validate.HTML}}</div>
+	{{end}}
+</div>
+
+{{template "_backend_bottom.gohtml" .}}
