+++ tpl/settings_purge.gohtml
@@ -1,0 +1,23 @@
+{{template "_backend_top.gohtml" .}}
+
+{{template "_settings_nav.gohtml" .}}
+
+<h2 id="purge">Purge</h2>
+
+<p>Remove all instances of a page.</p>
+
+<p>Matches are case insensitive. Supports <code>%</code> as a wildcard; e.g.
+	<code>/page%.html</code> matches everything starting with <code>/page</code> and
+	ending with <code>.html</code>. <code>_</code> matches any character; e.g.
+	<code>_.html</code> matches <code>a.html</code> and <code>b.html</code>. Use
+	<code>\%</code> and <code>\_</code> for the literal characters without special
+	meaning.</p>
+
+<form method="get" action="/settings/purge/confirm">
+	<input type="text" name="path" placeholder="Path" required autocomplete="off">
+	<button type="submit">Purge</button>
+	<span>You will see a preview of matches before anything is deleted</span><br>
+	<label><input type="checkbox" name="match-title"> Match title as well</label>
+</form>
+
+{{template "_backend_bottom.gohtml" .}}
