+++ tpl/settings_purge_confirm.gohtml
@@ -13,7 +13,7 @@
 		</tbody>
 	</table>
 
-	<form method="post">
+	<form method="post" action="/settings/purge">
 		<input type="hidden" name="csrf" value="{{.User.CSRFToken}}">
 		<input type="hidden" name="paths" value="{{range $s := .List}}{{$s.PathID}},{{end}}">
 		<button>Yes, purge them all!</button>
